Commit auto-generated files
Checking out `main` and running `make` produces these updates. I'm not familiar with the devbox setup, but `devbox.lock` was already committed so I assume we would want to commit changes to it. The `gradle.properties` change seems like it could be gitignored? I've committed it here but I can add it to the gitignore if that makes more sense